--- Mariya/config.py
@@ -1,11 +1,0 @@
-import os
-
-time_sleep = 3
-
-path_to_home_dir = os.getcwd()                                  #пусть распложения main
-path_to_data_dir = path_to_home_dir + '\\data'                  #путь к папке
-path_to_data = path_to_data_dir + '\\dataNEW.xlsx'              #путь к файлу DATA 
-sourses = "\\sourses"                                           #?????????
-path_to_loads_dir = path_to_data_dir + '\\loads'
-path_to_json = path_to_data_dir + '\\json.json'
-final_exel_file = path_to_home_dir + '\\finalfile.xlsx'
