+++ Mariya/main.py
@@ -1,11 +1,12 @@
+# -*- coding: utf-8 -*-
 from time import sleep
 import sys
 from PySide6.QtWidgets import QApplication, QMainWindow
-from ui import Ui_Dialog
+from ui_module import Ui_Dialog
 import time 
-
-
-start = time.time()
+from report_module import read_exel_inn, writer_a_report_file, append_in_data
+from config_module import path_to_data, path_to_final_exel_file
+from requests_module import get_id, get_organization, report_generation
 
 
 def progress_bar(steps: int, max_lenght: int = 50) -> None:
@@ -24,6 +25,29 @@ def progress_bar(steps: int, max_lenght: int = 50) -> None:
 
 #progress_bar(100)
 
+def main_Function():
+	try:
+		data = read_exel_inn(path_to_data)
+		for inn in data:
+			#append_in_data(report_generation(get_organization(get_id(inn)), inn))
+			#writer_a_report_file(report_generation(get_organization(get_id(inn)), inn))
+			report_generation(get_organization(get_id(inn)), inn)
+			#report_generation(get_organization(get_id(inn)), inn)
+			#time.sleep(3)
+	except Exception as ex:
+		print(f'Ошибка: {ex}')
+main_Function()
+
+#def example(inn):
+#	writer_a_report_file(report_generation(get_organization(get_id(inn)), inn))
+#	return("Example >>> Complite")
+#
+#data = read_exel_inn(path_to_data)
+#for inn in data:
+#	example(inn)
+#	next(data)
+
+
 
 class App(QMainWindow):
 	def __init__(self):
@@ -32,14 +56,21 @@ class App(QMainWindow):
 		self.ui.setupUi(self)
 	
 
-end = time.time()
-total_time = (end - start)/60
-print(f'Время выполнения: {total_time} мин.')
-
 if __name__ == '__main__':
-	
+	start = time.time()
+
 	app = QApplication(sys.argv)
 	window = App()
 	window.show()
 
-	sys.exit(app.exec())
+
+	end = time.time()
+	total_time = (end - start)/60
+	print(f'Время выполнения: {total_time} мин.')
+
+
+	sys.exit(app.exec())	
+
+
+
+
